[server] ArmZabbixAPI,ConfigManager: Do not get old events at first by default.
When adding a new zabbix server, Hatohol tried to get all events
from the server.  It may take very long time if the server has
accumulated tons of events.  This is very harmful because it
places a high load on the monitoring server.

Do not load old events by default unless --load-old-events option
is specified.

Signed-off-by: YOSHIFUJI Hideaki <hideaki.yoshifuji@miraclelinux.com>