[hatohol-db-initiator] Add an option to be carried out w/o config. file.
Current implementation cannot be used when the config. file is
not correctly installed. However, mainly developers sometimes
want to use this script w/o installation. This patch enables it.

As for the DB user name and the password for Hatohol server
the default values of the previous version (user: hatohol,
password: hatohol), is automatically used.
i.e. The user can omit --hatohol-db-user and --hatohol-db-password.