[DBTablesUser] Use DBTermCStringProver to make a part of an SQL statement.
The instance name 'rhs' means Right Hand Side of a condition
equation in an SQL statement. It also aims to avoid readers
from being confused. In this project, instances of DBTables
and the child classes have been named dbXXXX such as
dbMonitoring and dbHost.

The name 'rhs' seems to be a too short and meaningless.  However,
lines to generate an SQL statement tend to be a long. So I believe
the short name makes the code simple after all.