[DBTablesMonitoring] Change getLastEventId() to getMaxEventId().
In a recent commit, getLastEventId() had changed to use cast().
in order to get the maximum number from the ones saved as a string.
However, it cannot utilize indexes. So it degrades the performance.

This patch just reverts it and uses max() again as with the code before.
Instead, the events have to be saved with the same digit number.
For Zabbix events, this has already been done in the previous commit.