hap2: Improvement the recover mechanism (#1366)
This patch devides constructor into two phases. One is, of course,
__init__() which is called when the object is created. In it,
we just put liens that don't raise exceptions. And setup() is
added as a bottom half of the constructor, which may cause
exceptions. This mechanism makes sure that the object is
certainly created and it enables calling destroy method of
the object if the setup() raises.

With the mechanism, I fixed #1366 by destroying object that
raised an exception.