Terminate the program if any problem of rabbitmq happen. (#1440)
When the problem happens in the call of pika, a client library for
RabbitMQ, an exception is raised. Then, it is caught in
StandardHap.__call__(). However, some situation this kind of
the problem happens needs to restart of the process in order to
initialize pika's internal state.

This patch terminates the process by making Signal with a critical
flag raise.