[server][test] Compare the return value with sorting.
The order of a returned event list by getEventInfoList() is
undetermined and depends on the used storage engine.
Thus comparision between the output and the expectation should be
done after sorting.

Actually, some tests such as test_getEventsSelectByServersAndHosts()
fail w/o this fix when MEMORY storage engine is used as.

  $ HATOHOL_MYSQL_ENGINE_MEMORY=1 MLPL_LOGGER_LEVEL=WARN ./run-test.sh -t /Monitoring/