[server][FaceRest] New classes: RestResourceStatic/MemberHandler.
In the previous implementation, ResourceHandler has both variables for
a static handler and a member handler, although only one of them
is used. And it has conditional branch in handle().

This patch adds the specialized class for them and makes the code
more simple with the minimum impact to the existing code.