[test] Terminate launced Hatohol server and manage.py.
test/run-client-tesst.sh launches Hatohol server and manage.py,
although it never terminate them.
It is no problem unless this script is used in Travis CI, because
the new Hatohol server and manange.py is not required and launched.
However, on developemnet, some situation need to terminate them
in order to run the script again.