[hathol-db-initiator] Don't setup HAPI1 plugins if not used. (#1626)
Previously the SQL files for setting up HAPI1 plugins such as
server-type-hapi-zabbix.sql are loaded by hatohol-db-initiator
even if QPID development package is unavaialble on the build,
which means the HAPI1 framework cannot be used.

This patch makes hatohol-db-initiator load the file only when
the HAPI1 framework is enabled (QPID development package is given
in the compile time).