[client] Remove affected styles
In the previous version following styles are applied to all ul &
li elements. It breaks the layout of navbar.

  display: inline-block;
  vertical-align: middle;

The above style is introduces at 29cac9bf to align "Latest events"
button and pager buttons. But it's not needed anymore because
"Latest events" button has been moved to other place.