[WebUI] Show an error message when Hatohol server is not working (#1581).
The previous implementation never show any message and doesn't
happen anything after the login button is puressed when the
Hatohol server is down.

Although it has the basic mechasim to show the message.
However, it's never been performed due to the exception at
$.parseJSON() with an empty data.

However, parsing the response data has been meaningless,
because parseLoginResultWithError() is called only when
the HTTP respose is NOT 200 (OK).

This patch simply show an error message when this method
is called back.