[WebUI] Redirect the top page to ajax_dashboard. (#1761)
Previous implementaion returns the content that is the same
as ajax_dashboard when the top page is requested. In this
case, a browser looks that the url /hatohol has the contents.
And if the login dialog is shown at that time, the JavaScript
code saves the session ID in the cookie for the page.
with the path: '/'.

However, the login dialog is shown at other pages such as
/hatohol/ajax_events, the path in the cookie is '/hatohol'.
Thus the two cookies are saved and the wrong (expired) one
may be used.

This patch ensures that all contents (pages) of WebUI are
under /hatohol.