[WebUI] Redirect the top page to ajax_dashboard. (#1761)
Previous implementaion returns the content of ajax_dashboard
when the result is for the top page (e.g. /hatohol).

The browser looks that the page /hatohol has the contents.
And if the login dialog is shown at the time, the JavaScript
code save the session ID in the cookie for the page.
i.e. the path is '/'.

However, the login dialog is shown at other pages such as
/hatohol/ajax_events, the path for the cookie is '/hatohol'.
Thus the two cookies are saved and the wrong one may be used.

This patch ensure that the login is always done under /hatohol.