Accept and stores multiple itemGroups (#1765).
Previosly, the HAPI2.0 interface implementation of Hatohol server
has accepted only the first item group (which calls it application in
ZABBIX) and drops the followed ones. This patch enables multiple
item groups to be handled properly.

To realize this feature, this patch adds the followings.

- adds a new table: item_categories instead of items.item_group_name
  which can have only one group name.

- 'itemGroupName' and 'applications' in FaceREST request are replaced
  'itemGroupNames' and 'ItemGroups' respectively for improving readability.

Note: The item groups (itemGroups) is called item categroy (itemCategory)
in Hatohol sever code to avoid confusion with the class ItemGroup already
existing, which is a generic container for ItemData.