Accept and stores multiple itemGroups (#1765).
Previosly, the HAPI2.0 interface of Hatohol server only accepts
the first item group and drops the followed ones.
This patch newly adds the tables for the item groups.

Note: Around interfaces such as HAPI2.0 and FaceREST, these
are called item group. However, in the hatohol server code,
they are named item category in order to avoid the confusion
with the class ItemGroup, which is the contaner for ItemData.