[tset][TriggersView] Fix an issue that a test fails on certain environments
The test 'With a problem trigger' fails on certain environments.

The cause is that the test expects a fixed HTML string although
it's modified by jQuery dynamically.
To resolve it, we assert only existence of "severity5" class and
backgroud color in this test. Because other parameters are tested
in other tests, we don't need test them in this test.

Fix #2009