[Hap][test] Fix the broken test.
The test has replaced sys.argv to hide some unittest arguments
and use a transporter for test (EzTransporter).

A recent chagange for ConfigFileParser added the partial parse
of the argument in the constructor of StandardHap. It is done
before the replacement of sys.argv. It may cause an unexpected
parse result.

This patch replaces sys.argv before it's parsed.