[WebUI][latest_view] Show items on explicit update operation.
Current implementation gets items without any condition. Althogh
the nubmer of items to be obtained is limited, this query might
make the Hatohol server heavy and take a while.

So with this patch, items aren't loaded automatically. By explit
update is need to load items typically after the query conditions
are set.